backend/requirements.txt
pandas>=3.0.1
numpy>=2.4.3
scikit-learn>=1.8.0
joblib>=1.5.3
xgboost>=3.2.0
matplotlib>=3.10.8
seaborn>=0.13.2



dotenv>=0.9.9
torch>=2.10.0
langchain>=1.2.13
langchain_openai>=1.1.12
langchain_ollama>=1.0.1
langgraph>=1.1.3
langgraph_checkpoint_sqlite>=3.0.3
requests_cache>=1.3.1
indictranstoolkit>=1.1.1
transformers>=4.45.2
tokenizers>=0.20.3
huggingface-hub>=0.23.2
sentencepiece>=0.2.1
